+++ services/monitor/src/ChainMonitor.ts
@@ -16,6 +16,7 @@ import type {
 } from "./types";
 import PendingContract from "./PendingContract";
 import type { Logger } from "winston";
+import type SimilarityVerificationClient from "./SimilarityVerificationClient";
 
 function createsContract(tx: TransactionResponse): boolean {
   return !tx.to;
@@ -31,6 +32,7 @@ export default class ChainMonitor extends EventEmitter {
   private sourceFetchers: KnownDecentralizedStorageFetchers;
   private sourcifyServerURLs: string[];
   private sourcifyRequestOptions: SourcifyRequestOptions;
+  private similarityVerificationClient: SimilarityVerificationClient;
 
   private chainLogger: Logger;
   private startBlock?: number;
@@ -47,10 +49,12 @@ export default class ChainMonitor extends EventEmitter {
     sourcifyChain: SourcifyChain,
     sourceFetchers: KnownDecentralizedStorageFetchers,
     monitorConfig: MonitorConfig,
+    similarityVerificationClient: SimilarityVerificationClient,
   ) {
     super();
     this.sourcifyChain = sourcifyChain;
     this.sourceFetchers = sourceFetchers; // TODO: handle multipe
+    this.similarityVerificationClient = similarityVerificationClient;
     this.chainLogger = logger.child({
       moduleName: "ChainMonitor #" + this.sourcifyChain.chainId,
       chainId: this.sourcifyChain.chainId,
@@ -286,6 +290,11 @@ export default class ChainMonitor extends EventEmitter {
           address,
           origin: metadataHash.origin,
         });
+        this.similarityVerificationClient.trigger(
+          this.sourcifyChain.chainId,
+          address,
+          creatorTxHash,
+        );
         return;
       }
 
@@ -300,6 +309,11 @@ export default class ChainMonitor extends EventEmitter {
         await pendingContract.assemble();
       } catch (err: any) {
         this.chainLogger.info("Couldn't assemble contract", { address, err });
+        this.similarityVerificationClient.trigger(
+          this.sourcifyChain.chainId,
+          address,
+          creatorTxHash,
+        );
         return;
       }
       if (!this.isEmpty(pendingContract.pendingSources)) {
@@ -307,6 +321,11 @@ export default class ChainMonitor extends EventEmitter {
           address: pendingContract.address,
           pendingSources: pendingContract.pendingSources,
         });
+        this.similarityVerificationClient.trigger(
+          this.sourcifyChain.chainId,
+          address,
+          creatorTxHash,
+        );
         return;
       }