+++ services/server/src/server/services/storageServices/AbstractDatabaseService.ts
@@ -5,6 +5,7 @@ import { bytesFromString } from "../utils/database-util";
 import type { DatabaseOptions } from "../utils/Database";
 import { Database } from "../utils/Database";
 import type { PoolClient, QueryResult } from "pg";
+import { ConflictError } from "../../../common/errors/ConflictError";
 
 export default abstract class AbstractDatabaseService {
   public database: Database;
@@ -215,6 +216,9 @@ export default abstract class AbstractDatabaseService {
 
       return verifiedContractInsertResult.rows[0].id;
     } catch (e) {
+      if (e instanceof ConflictError) {
+        throw e;
+      }
       throw new Error(
         `cannot update verified_contract address=0x${databaseColumns.contractDeployment.address.toString("hex")} chainId=${databaseColumns.contractDeployment.chain_id}\n${e}`,
       );