+++ services/server/src/server/services/storageServices/AllianceDatabaseService.ts
@@ -3,6 +3,7 @@ import AbstractDatabaseService from "./AbstractDatabaseService";
 import type { WStorageService } from "../StorageService";
 import type { VerificationExport } from "@ethereum-sourcify/lib-sourcify";
 import { WStorageIdentifiers } from "./identifiers";
+import { ConflictError } from "../../../common/errors/ConflictError";
 
 export class AllianceDatabaseService
   extends AbstractDatabaseService
@@ -21,18 +22,25 @@ export class AllianceDatabaseService
           transactionPoolClient,
         );
       });
+      logger.info("Stored to AllianceDatabase", {
+        name: verification.compilation.compilationTarget.name,
+        address: verification.address,
+        chainId: verification.chainId,
+        runtimeMatch: verification.status.runtimeMatch,
+        creationMatch: verification.status.creationMatch,
+      });
     } catch (error) {
+      if (error instanceof ConflictError) {
+        logger.warn("Contract already exists in AllianceDatabase", {
+          name: verification.compilation.compilationTarget.name,
+          address: verification.address,
+        });
+        throw error;
+      }
       logger.error("Error storing verification", {
         error: error,
       });
       throw error;
     }
-    logger.info("Stored to AllianceDatabase", {
-      name: verification.compilation.compilationTarget.name,
-      address: verification.address,
-      chainId: verification.chainId,
-      runtimeMatch: verification.status.runtimeMatch,
-      creationMatch: verification.status.creationMatch,
-    });
   }
 }